# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        }
    },
    "paths": {
...
        "/api/v1/base/tag-manages": {
            "get": {
                "description": "查询标签列表",
                "tags": [
                    "TagManage"
                ],
                "summary": "查询标签列表",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "是否仅查询count",
                        "name": "-",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "当前页",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标签名称",
                        "name": "key_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页大小",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "模糊搜索(模糊查询:标签名称、标签值)",
                        "name": "query_value",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标签类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标签值",
                        "name": "value",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.ListResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            },
            "post": {
                "description": "创建标签",
                "tags": [
                    "TagManage"
                ],
                "summary": "创建标签",
                "parameters": [
                    {
                        "description": "创建标签参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.TagManage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.IDResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
        "/api/v1/base/tag-manages/:id": {
            "get": {
                "description": "查询指定标签详情",
                "tags": [
                    "TagManage"
                ],
                "summary": "查询指定标签详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "标签ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.TagManage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            },
            "put": {
                "description": "更新标签信息",
                "tags": [
                    "TagManage"
                ],
                "summary": "更新标签信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "标签ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新标签信息参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.TagManage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除标签信息",
                "tags": [
                    "TagManage"
                ],
                "summary": "删除标签信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "标签ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.IDResult": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
        "common.PaginationResult": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "ginx.ListResult": {
            "type": "object",
            "properties": {
                "list": {},
                "pagination": {
                    "$ref": "#/definitions/common.PaginationResult"
                }
            }
        },
        "ginx.OperateResult": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
...
        "schema.TagManage": {
            "type": "object",
            "required": [
                "type",
                "value"
            ],
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "integer"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "integer"
                },
                "key_name": {
                    "description": "标签名称",
                    "type": "string"
                },
                "type": {
                    "description": "标签类型",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "value": {
                    "description": "标签值",
                    "type": "string"
                }
            }
        },
...
    }
}